requirements.txt
fastapi
uvicorn
sentence-transformers>=2.5.1
torch
accelerate
transformers
numpy
pydantic
psycopg2-binary
pgvector
beautifulsoup4
lxml
requests
